feat: Updated page layout structure
