+++ web/app/(frontend)/(logged-in)/[assessmentGroupId]/assessments/[assessmentId]/[partName]/[sectionId]/[attributeId]/layout.tsx
@@ -1,0 +1,56 @@
+import { 
+    fetchAssessmentType, 
+    fetchAssessment,
+    fetchPart,
+    fetchSection,
+    fetchAttribute
+  } from "../../../../../../utils/dataFetchers"
+  import NotFound from "@/app/(frontend)/components/notFound"
+  
+  export default async function RootLayout({
+    children,
+    params
+  }: Readonly<{
+    children: React.ReactNode,
+    params: { 
+      assessmentGroupId: string,
+      assessmentId: string,
+      partName: string,
+      sectionId: string,
+      attributeId: string
+    }
+  }>) {
+    const assessmentType = await fetchAssessmentType(params.assessmentGroupId)
+    const assessment = await fetchAssessment(params.assessmentId)
+    const part = await fetchPart(params.assessmentGroupId, params.partName)
+    const section = await fetchSection(params.sectionId)
+    const attribute = await fetchAttribute(params.attributeId)
+  
+    if (assessmentType && assessment && part && section) {
+      const links = [
+        {
+            url: `/${assessmentType.id}/assessments`, 
+            name: assessmentType.name
+        },
+        {
+            url: `/${assessmentType.id}/assessments/${assessment.id}`, 
+            name: assessment.name
+        },
+        {
+            url: `/${assessmentType.id}/assessments/${assessment.id}/${part.name}`, 
+            name: part.name
+        },
+        {
+            url: `/${assessmentType.id}/assessments/${assessment.id}/${part.name}/${section.id}`,
+            name: `${section.id.toString().toUpperCase()}. ${section.name}`
+          }
+      ]
+      if (attribute) {
+        return (
+            {children}
+        )
+      }
+      return <NotFound links={links} pageType="attribute" />
+    }
+    
+  }
