+++ web/app/(frontend)/(logged-in)/[assessmentGroupId]/assessments/[assessmentId]/[partName]/layout.tsx
@@ -1,0 +1,42 @@
+import { 
+    fetchAssessmentType, 
+    fetchAssessment,
+    fetchPart
+  } from "../../../../utils/dataFetchers"
+  import NotFound from "@/app/(frontend)/components/notFound"
+  
+  export default async function RootLayout({
+    children,
+    params
+  }: Readonly<{
+    children: React.ReactNode,
+    params: { 
+      assessmentGroupId: string,
+      assessmentId: string,
+      partName: string
+    }
+  }>) {
+    const assessmentType = await fetchAssessmentType(params.assessmentGroupId)
+    const assessment = await fetchAssessment(params.assessmentId)
+    const part = await fetchPart(params.assessmentGroupId, params.partName)
+  
+    if (assessmentType && assessment) {
+      const links = [
+        {
+          url: `/${assessmentType.id}/assessments`, 
+          name: assessmentType.name
+        },
+        {
+            url: `/${assessmentType.id}/assessments/${assessment.id}`, 
+            name: assessment.name
+          },
+      ]
+      if (part) {
+        return (
+            {children}
+        )
+      }
+      return <NotFound links={links} pageType="part" />
+    }
+    
+  }
