+++ web/app/(frontend)/(logged-in)/[assessmentGroupId]/assessments/[assessmentId]/[partName]/page.tsx
@@ -7,6 +7,7 @@ import {
 import { auth } from "@/auth"
 import { isParticipantForAssessment, viewableResponses } from "../../../../utils/permissions"
 
+import Breadcrumbs from "@/app/(frontend)/components/breadcrumbs"
 import PartContent from "./part"
 
 export default async function Page({ params }: Readonly<{ params: { 
@@ -24,16 +25,45 @@ export default async function Page({ params }: Readonly<{ params: {
   const isParticipant = isParticipantForAssessment(session, params.assessmentId)
   const numAssessmentUsers = isParticipant ? 1 : assessmentParticipants.length
   const userResponses = await viewableResponses(session, params.assessmentId)
-  
-  return (
-    <div className="flex h-full w-full flex-col items-center justify-start pt-3 pb-10">
-      <PartContent 
-        assessment={assessment} 
-        assessmentType={assessmentType} 
-        part={part} 
-        numAssessmentUsers={numAssessmentUsers}
-        userResponses={userResponses} 
-      />
-    </div>
-  )
+
+  if (assessmentType && assessment) {
+    const links = [
+      {
+        url: `/${assessmentType.id}/assessments`, 
+        name: assessmentType.name
+      },
+      {
+        url: `/${assessmentType.id}/assessments/${assessment.id}`,
+        name: assessment.name
+      }
+    ]
+    if (part) {
+      return (
+        <div className="w-full max-w-4xl mx-auto">
+          <section className="mb-8">
+            <div className="space-y-4 max-lg:ml-2">
+              <Breadcrumbs links={links} currentPage={part.name} />
+              <div className="flex flex-row justify-between">
+                <div className="space-y-2">
+                    <h1 className="text-3xl font-bold tracking-tighter">{part.name}</h1>
+                    <p className="text-sm text-muted-foreground dark:text-indigo-300/80">
+                      {part.description}
+                    </p>
+                </div>
+              </div>
+            </div>
+          </section>
+          <section className="mb-16">
+            <PartContent 
+              assessment={assessment} 
+              assessmentType={assessmentType} 
+              part={part} 
+              numAssessmentUsers={numAssessmentUsers}
+              userResponses={userResponses} 
+            />
+          </section>
+        </div>
+      )
+    }
+  }
 }
