+++ web/app/(frontend)/(logged-in)/[assessmentGroupId]/assessments/[assessmentId]/page.tsx
@@ -13,6 +13,9 @@ import {
   viewableResponses 
 } from "../../../utils/permissions"
 
+import { Button } from "@/components/ui/button"
+import Link from "next/link"
+import Breadcrumbs from "@/app/(frontend)/components/breadcrumbs"
 import AssessmentContent from "./assessment"
 
 export default async function Page({ params }: Readonly<{ params: { assessmentGroupId: string, assessmentId: string } }>) {
@@ -35,17 +38,50 @@ export default async function Page({ params }: Readonly<{ params: { assessmentGr
     isManagerForCollection(session, assessment?.assessmentCollectionId) || 
     isLeadForAssessment(session, params.assessmentId) ||
     permissions?.find(permission => permission.name === "Edit assessments") !== undefined
-  
-  return (
-    <div className="flex h-full w-full flex-col items-center justify-start pt-3 pb-10">
-      <AssessmentContent 
-        assessment={assessment} 
-        assessmentType={assessmentType} 
-        parts={parts} 
-        numAssessmentUsers={numAssessmentUsers}
-        userResponses={userResponses} 
-        canEdit={canEdit} 
-      />
-    </div>
-  )
+
+    if(assessmentType && assessment) {
+      const links = [
+        {
+          url: `/${assessmentType.id}/assessments`, 
+          name: assessmentType.name
+        },
+      ]
+      return (
+        <div className="w-full max-w-4xl mx-auto">
+          <section className="mb-8">
+            <div className="space-y-4 max-lg:ml-2">
+              <Breadcrumbs links={links} currentPage={assessment.name} />
+              <div className="flex flex-row justify-between">
+                <div className="space-y-2">
+                  <h1 className="text-3xl font-bold tracking-tighter">{assessment.name}</h1>
+                  <p className="text-sm text-muted-foreground dark:text-indigo-300/80">
+                    {assessment.description}
+                  </p>
+                </div>
+                {/* TODO: Edit assessment functionality */}
+                {canEdit && <div>
+                  <Link
+                    href={`/${assessmentType.id}/assessments/${assessment.id}/edit-assessment`}
+                    prefetch={false}
+                  >
+                    <Button>
+                      Edit Assessment
+                    </Button>
+                  </Link>
+                </div>}
+              </div>
+            </div>
+          </section>
+          <section className="mb-16">
+            <AssessmentContent 
+              assessment={assessment} 
+              assessmentType={assessmentType} 
+              parts={parts} 
+              numAssessmentUsers={numAssessmentUsers}
+              userResponses={userResponses}
+            />
+          </section>
+        </div>
+      )
+    }
 }
