+++ web/app/(frontend)/(logged-in)/[assessmentGroupId]/home.tsx
@@ -100,17 +100,6 @@ export default function Home({
       </div>
     )
   }
-  return (
-    <div className="w-full max-w-4xl mx-auto">
-        <section className="mb-8">
-            <div className="space-y-8 max-lg:ml-2">
-                <p className="text-md text-muted-foreground dark:text-indigo-300/80">
-                    The assessment type could not be found.
-                </p>
-            </div>
-        </section>
-    </div>
-  )
 }
 
 function AssessmentCard({
