+++ web/app/(frontend)/(logged-in)/[assessmentGroupId]/page.tsx
@@ -13,14 +13,12 @@ export default async function Page({ params }: Readonly<{ params: { assessmentGr
   const userResponses = await fetchUserResponses(session?.user?.id)
   
   return (
-    <div className="flex h-full flex-col items-center justify-start p-2 sm:p-10">
-      <Home 
-        assessmentType={assessmentType} 
-        assessments={assessments} 
-        parts={parts} 
-        userResponses={userResponses} 
-        session={session} 
-      />
-    </div>
+    <Home 
+      assessmentType={assessmentType} 
+      assessments={assessments} 
+      parts={parts} 
+      userResponses={userResponses} 
+      session={session} 
+    />
   )
 }
