+++ web/app/(frontend)/(logged-in)/[assessmentGroupId]/users/page.tsx
@@ -10,9 +10,9 @@ export default async function Page({ params }: Readonly<{ params: { assessmentGr
   const assessmentType = await fetchAssessmentType(params.assessmentGroupId)
   const assessments = await viewableAssessments(session, params.assessmentGroupId)
 
-  return (
-    <div className="flex h-full flex-col items-center justify-start p-2 sm:p-10">
+  if (assessmentType) {
+    return (
       <DataTable assessments={assessments} assessmentType={assessmentType} />
-    </div>
-  )
+    )
+  }
 }
